rename PluginHost to HookHost
